Moved stage and characters lists into separate file
