+++ .gitignore
@@ -1,1 +1,2 @@
 *.json
+__pycache__
