+++ __init__.py
@@ -1,0 +1,4 @@
+# __init__.py
+from .game_data import stages, characters
+
+__all__ = ["stages", "characters"]
